requirements-dev.txt
-r requirements.txt

Flask-WebTest==0.0.7
pytest-cov==1.8.1
coveralls==0.5
